Split source code into multiple files
While having a single `.cpp` is a nice thing for minimal application
that just renders Markdown, extending the applications features would
become messy unless I start splitting the source code to multiple files.
